backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
chromadb
requests
beautifulsoup4
pandas
numpy
pypdf
langchain